Revert "temp" 놈체크마치고 나니까 general command 작동이 안 돼서 취소함.
This reverts commit 17aa6da34685f6a961afe83e314c8cabb37f93e8.
