+++ Makefile
@@ -6,7 +6,7 @@
 #    By: han-yeseul <han-yeseul@student.42.fr>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/07/15 21:26:09 by minkyeki          #+#    #+#              #
-#    Updated: 2022/07/28 17:14:11 by han-yeseul       ###   ########.fr        #
+#    Updated: 2022/07/28 16:11:26 by han-yeseul       ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -14,9 +14,9 @@ NAME			= minishell
 CC				= cc
 
 # 				#-g3d 옵션 붙이면 오류가 뜬 파일 위치와 줄번호까지 뜸
-CFLAGS		= -Wall -Wextra -Werror -g3 -fsanitize=address
+# CFLAGS		= -Wall -Wextra -Werror -g3 -fsanitize=address
 
-# CFLAGS			= -Wall -Wextra -Werror
+CFLAGS			= -Wall -Wextra -Werror
 INCLUDE			= include
 
 # GNU readline Flag for MAC
@@ -44,7 +44,7 @@ LEXER_SRC		= token_create token_modify \
 				  scanner_helper scanner_get_token_1 scanner_get_token_2
 
 PARSER_SRC		= parse_tree parse_tree_utils
-EXECUTER_SRC	= executer token_expand token_expand_utils get_full_path subshell \
+EXECUTER_SRC	= executer token_expand get_full_path subshell \
 				  redirection redirection_utils exec_command exec_command_utils
 
 BUILTIN_SRC		= environ environ_utils environ_helper cd echo pwd exit
